0407 아이템 추가 및 세이브/로드/창고 정리 완성

+++ Assets/JB/ItemGenerateManager.cs
@@ -28,7 +28,7 @@ public class ItemGenerateManager : MonoBehaviour
         }
     }
 
-    public void GenItem(ushort id, sbyte amount, short duration, byte grade)
+    public void GenItem(ushort id, short amount, short duration, byte grade)
     {
         ItemDataContainer item = ObjectPool.GetObject();
         item.SetData(new ItemObjectData(id, amount, duration, grade));
